# openapi.yaml
openapi: 3.0.0
info:
  title: Atende API
  version: 1.0.0
  description: API for the Atende application
paths:
...
  /system/{sigla}/{type}:
    get:
      summary: Get system image (icon or logo)
      parameters:
        - in: path
          name: sigla
          required: true
          schema:
            type: string
          description: System acronym/sigla
        - in: path
          name: type
          required: true
          schema:
            type: string
            enum: [icon, logo]
          description: Type of image to retrieve (icon or logo)
      responses:
        '200':
          description: System image retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  image:
                    type: string
                    description: Base64 encoded image
                  type:
                    type: string
                    description: Image MIME type
        '400':
          description: Invalid image type
        '404':
          description: System or image not found
...
  /api/cruds:
    get:
      summary: Get all CRUDs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of CRUDs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Crud'
    post:
      summary: Create a new CRUD
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrudInput'
      responses:
        '201':
          description: CRUD created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crud'
        '400':
          description: Bad request - Invalid configuration or duplicate name
  /api/cruds/{id}:
    get:
      summary: Get a CRUD by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A single CRUD
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crud'
        '404':
          description: CRUD not found
    put:
      summary: Update a CRUD by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrudInput'
      responses:
        '200':
          description: CRUD updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crud'
        '400':
          description: Bad request - Invalid configuration or duplicate name
        '404':
          description: CRUD not found
    delete:
      summary: Delete a CRUD by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: CRUD deleted
        '404':
          description: CRUD not found
...
  /api/mcp:
    post:
      summary: Model Context Protocol (MCP) server endpoint
      description: |
        Servidor MCP que expõe ferramentas disponíveis para modelos de IA.
        Suporta métodos JSON-RPC 2.0: initialize, tools/list, tools/call
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - jsonrpc
                - method
              properties:
                jsonrpc:
                  type: string
                  enum: ['2.0']
                  description: Versão do protocolo JSON-RPC
                method:
                  type: string
                  enum: ['initialize', 'tools/list', 'tools/call']
                  description: Método MCP a ser executado
                params:
                  type: object
                  description: Parâmetros do método (varia conforme o método)
                  additionalProperties: true
                id:
                  oneOf:
                    - type: string
                    - type: integer
                  description: ID da requisição para correlação de resposta (opcional, pode ser omitido para null)
      responses:
        '200':
          description: Resposta MCP bem-sucedida
          content:
            application/json:
              schema:
                type: object
                properties:
                  jsonrpc:
                    type: string
                    enum: ['2.0']
                  result:
                    type: object
                    description: Resultado do método (varia conforme o método)
                    additionalProperties: true
                  error:
                    type: object
                    description: Erro se houver
                    properties:
                      code:
                        type: integer
                      message:
                        type: string
                      data:
                        type: object
                    additionalProperties: true
                  id:
                    oneOf:
                      - type: string
                      - type: integer
                    description: ID da requisição (opcional, pode ser null)
        '400':
          description: Requisição inválida
        '500':
          description: Erro interno do servidor
  /api/chatia:
    post:
      summary: Chat with AI assistant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  description: User message to the AI
                conversation:
                  type: array
                  description: Previous conversation messages
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                        enum: [user, assistant, system]
                      content:
                        type: string
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: AI response message
                  functionCalls:
                    type: array
                    description: Functions executed by the AI
                    items:
                      type: object
                      properties:
                        function:
                          type: string
                        params:
                          type: object
                        result:
                          type: object
        '401':
          description: Unauthorized
        '500':
          description: Server error
...
  /api/models:
    get:
      summary: Get all models
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelInfo'
    post:
      summary: Create a new model
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelInput'
      responses:
        '201':
          description: Model created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
        '400':
          description: Bad request
        '403':
          description: Forbidden
  /api/models/{name}:
    get:
      summary: Get a model by name
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
          description: Model name
      responses:
        '200':
          description: A single model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
        '404':
          description: Model not found
    put:
      summary: Update a model
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
          description: Model name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelInput'
      responses:
        '200':
          description: Model updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  content:
                    type: string
        '404':
          description: Model not found
...
components:
...
  schemas:
...
    Crud:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          description: Unique identifier for the CRUD
        title:
          type: string
          description: Display title
        icon:
          type: string
          description: Material icon name
        resource:
          type: string
          description: Resource name for permissions
        endpoint:
          type: string
          description: API endpoint path
        config:
          type: object
          description: JSON configuration object
          additionalProperties: true
        active:
          type: boolean
          description: Whether the CRUD is active
      required:
        - name
        - title
        - resource
        - endpoint
        - config
    CrudInput:
      type: object
      properties:
        name:
          type: string
          description: Unique identifier for the CRUD
        title:
          type: string
          description: Display title
        icon:
          type: string
          description: Material icon name
        resource:
          type: string
          description: Resource name for permissions
        endpoint:
          type: string
          description: API endpoint path
        config:
          type: object
          description: JSON configuration object
          additionalProperties: true
        active:
          type: boolean
          description: Whether the CRUD is active
      required:
        - name
        - title
        - resource
        - endpoint
        - config
    ModelInfo:
      type: object
      properties:
        name:
          type: string
        className:
          type: string
        file:
          type: string
        path:
          type: string
    Model:
      type: object
      properties:
        name:
          type: string
        className:
          type: string
        fields:
          type: array
          items:
            $ref: '#/components/schemas/Field'
        associations:
          type: array
          items:
            $ref: '#/components/schemas/Association'
        options:
          type: object
        rawContent:
          type: string
    ModelInput:
      type: object
      properties:
        name:
          type: string
        className:
          type: string
        fields:
          type: array
          items:
            $ref: '#/components/schemas/Field'
        associations:
          type: array
          items:
            $ref: '#/components/schemas/Association'
        options:
          type: object
          additionalProperties: true
      required:
        - name
        - className
      additionalProperties: true
    Field:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        allowNull:
          type: boolean
        primaryKey:
          type: boolean
        autoIncrement:
          type: boolean
        defaultValue:
          type: string
          description: Optional default value for the field
        unique:
          type: boolean
        references:
          type: object
          properties:
            model:
              type: string
            key:
              type: string
          additionalProperties: true
      additionalProperties: true
    Association:
      type: object
      properties:
        type:
          type: string
          enum: [belongsTo, hasMany, hasOne, belongsToMany]
        source:
          type: string
        target:
          type: string
        foreignKey:
          type: string
        otherKey:
          type: string
        through:
          type: string
      additionalProperties: true
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT